+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.0.1] - 2021-08-18
+### Added
+- Prefix for editor prefs
+### Changed
+- Package type from library to tool
+
 ## [1.0.0] - 2021-07-29
 ### Added
 - First public release